Move endpoint logic to separate files
