+++ app/models/rule.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class Rule(BaseModel):
+    name: str
+    group: str
+    operator: str
+    parenthesis: str
